+++ remote/a.json
@@ -1,1 +1,1 @@
-{"workingDirectory":[{"name":"b","state":"Modified","date":"2019-11-26 15:46:00"}],"stagingArea":[],"gitRepository":[],"logs":[{"id":"15h336d55f","message":"adf","date":"2019-11-26T15:45:50.610Z"}]}
+{"workingDirectory":[{"name":"b","state":"Modified","date":"2019-11-26 15:46:00"}],"stagingArea":[],"gitRepository":[],"logs":[]}
